models/requirements.txt
# ===================================================================
# CRITICAL: torch and dgl MUST be installed manually first
# using the specific, version-pinned commands for your CUDA version.
# See README for the correct pip commands.
#
# Example (CUDA 12.1):
# pip install torch==2.1.2+cu121 torchvision==0.16.2+cu121 torchaudio==2.1.2+cu121 --index-url https://download.pytorch.org/whl/cu121
# pip install dgl==2.2.1+cu121 -f https://data.dgl.ai/wheels/cu121/repo.html
# ===================================================================

# Core Data & ML Libraries
numpy==2.1.2
pandas==2.3.2
scikit-learn==1.7.1
scipy==1.15.3
torchdata==0.11.0

# Graph Learning
dgl==2.2.1+cu121   # install manually as per above

# Database Connector
neo4j==5.28.2

# Explainable AI
shap==0.48.0

# Utilities
tqdm==4.67.1
psutil==7.0.0
